requirements.txt
streamlit==1.37.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
pdfplumber==0.11.4
python-docx==1.1.2
requests==2.32.3
